First release of the thoery test site.
This version simply fetches all the questions from the resource (as of 05/06/2021 - total of 1802) as they are.
Since this specific dataset includes the HTML tags in each question's options by itself,
no further styling is done (except a little bit CSS). Therefore, the questions, with their corresponding
answers, are displayed as they were received from the resource.